+++ index.ts
@@ -1,11 +1,13 @@
 import TimeNotification from './functions/timeNotification'
 require('dotenv').config()
 const http = require('http');
+const mongoose = require('mongoose')
 
 const TelegramBot = require('node-telegram-bot-api');
 
 import { BotCommand } from './commandAPI'
 import * as botConfig from './botConfig'
+import * as nicknameCRUD from './MongoDB/CRUD/CustomNicknamesCRUD'
 
 const token = process.env.TOKEN;
 const bot = new TelegramBot(token, { polling: true });
@@ -23,7 +25,7 @@ loadCommands(`${__dirname}/functions/commands`);
 server.listen(process.env.PORT || 5000, () => {
     console.log(`Server running`);
 
-    // bot.sendMessage(process.env.CHAT_ID, 'Updated to version 4.1.5\nFixed:\n - Notification on 4:20\n- /next command\n\nIf you have any doubt, contribution or suggestion please feel free to hit me up in Telegram @suembra or create an issue in Github https://github.com/demo-hub/PinkGlobal420TelegramBot\nThank you kaidey for your contributions! You can find him in https://github.com/Kaidey\nIf you want to help please go on Github and star and/or watch the project. Maybe we can appear on Trending!\n\nBrace yourselves the next version is gonna be the version 4.2.0! Big things are coming!')
+    bot.sendMessage(process.env.CHAT_ID, 'Updated to version 4.2.0\nYou can see the changes in https://github.com/demo-hub/PinkGlobal420TelegramBot/blob/master/Changelog.md\n\nIf you have any doubt, contribution or suggestion please feel free to hit me up in Twitter or Telegram @suembra or create an issue in Github https://github.com/demo-hub/PinkGlobal420TelegramBot.\nThank you José Pinto (@JoseAoQuadrado) and Luis Azevedo (https://github.com/LuisAzeved) for your contributions!\nIf you want to help please go on Github and star and/or watch the project or follow me and José Pinto on Twitter @suembra and @JoseAoQuadrado')
 
     new TimeNotification().timeNotification(bot);
 
@@ -56,13 +58,32 @@ server.listen(process.env.PORT || 5000, () => {
 
 });
 
-function handleMentions(msg: any) {
+async function handleMentions(msg: any) {
+  var checkBotMention = new RegExp(/(.?)@PinkGlobal420Bot(.?)/);
 
-    var checkBotMention = new RegExp(/(.?)@PinkGlobal420Bot(.?)/);
+  const connector = mongoose.connect(botConfig.connectionString, {
+    useNewUrlParser: true
+  });
 
-    if (checkBotMention.test(msg.text)) {
-        bot.sendMessage(msg.chat.id, "Fazeçe irmão");
+  let customNick = await connector.then(async () => {
+    return nicknameCRUD.GetUserByID(msg.from.id);
+  });
+
+  if (checkBotMention.test(msg.text)) {
+    if (customNick == null) {
+      if (!msg.from.username) {
+        customNick = msg.from.first_name;
+      } else {
+        customNick = "@" + msg.from.username;
+      }
     }
+
+    var mention = `[${customNick}](tg://user?id=${msg.from.id})`;
+    bot.sendMessage(msg.chat.id, `Hi ${mention}\nFazeçe irmão`, {
+      parse_mode: "MarkdownV2"
+    });
+    // bot.sendMessage(msg.chat.id, "Fazeçe irmão");
+  }
 }
 
 function handleUserMessages(msg: any) {